requirements.txt
 # Simple Macro Agent Requirements
# Python 3.11+ compatible

# Core dependencies
python-dotenv==1.0.0
anthropic==0.18.1

# Data fetching
aiohttp==3.9.1
yfinance==0.2.33
pandas==2.1.4
numpy==1.26.2

# Async support
asyncio

# Email support (built-in)
# smtplib and email are part of Python standard library

# Optional for development
pytest==7.4.3
pytest-asyncio==0.21.1
